Rename tests/resources to tests/tests_resources for consistent module naming.
These files do not contain helper methods for resources.?; they provide
definitions for unit tests.  The convention followed currently then is to
exclude the prefix for the former, to prefix "tests_" in the case of the
latter, except for the base "testing" module and "main.c", and to add a prefix
of "test_" for files that contain declarations or definitions for unit tests.